requirements.reliai-adapter.txt
fastapi==0.115.6
uvicorn==0.30.6
httpx==0.27.2
opentelemetry-proto==1.25.0
